# implementation/src/main/java/com/assessment/authentication/model/RegisterAccountRequest.java
package com.assessment.authentication.model;

import lombok.Getter;
import lombok.Setter;
import lombok.ToString;

@Getter
@Setter
@ToString
public class RegisterAccountRequest {

    private String accountNumber;
    private String username;
    private String password;
}